+++ java-client/src/it/scala/InventoryIntegrationSpec.scala
@@ -1,0 +1,7 @@
+package sphere
+
+import org.scalatest._
+
+class InventoryIntegrationSpec extends WordSpec with MustMatchers {
+
+}